conflicts solve between variable types

+++ include/EncoderHandler.h
@@ -5,22 +5,21 @@
 #include <Encoder.h>
 
 // Encoder pins
-#define ENCODER_PIN_A 2  // CLK
-#define ENCODER_PIN_B 3  // DT
-#define ENCODER_BUTTON_PIN 4  // SW
+#define ENCODER_PIN_A 2      // CLK
+#define ENCODER_PIN_B 3      // DT
+#define ENCODER_BUTTON_PIN 4 // SW
 
-#define TEMP_STEP 1  // Step size for temperature adjustment
+#define TEMP_STEP 1 // Step size for temperature adjustment
 
 extern bool setMode;
 extern int currentOption;
-extern int set_T1, set_T2;
-
+extern double set_T1, set_T2;
 
 // Function declarations
-void setupEncoder();   // Initialize encoder
-void handleEncoder();  // Read encoder state
+void setupEncoder();    // Initialize encoder
+void handleEncoder();   // Read encoder state
 bool isButtonPressed(); // Check if the button is pressed
 
-extern int selectedThermocouple;  // T1 = 0, T2 = 1
+extern int selectedThermocouple; // T1 = 0, T2 = 1
 
 #endif
